requirements.txt
django-allauth==0.39.1
Django==2.2
django-filter==2.1.0
django-rest-auth==0.9.5
djangorestframework==3.9.2
pep8==1.7.1
pytz==2018.9
dnspython3==1.15.0
djongo==1.2.32
pymongo==3.7.2
gunicorn==19.9.0
whitenoise==4.1.2
sqlparse==0.2.4
django-heroku==0.3.1
django-reset-migrations==0.3.1
django-cors-headers==2.5.2
drf-nested-routers==0.91